refactor: extract Facade to its own class (#7884)
